Safari-611.1.10/LayoutTests/fast/multicol/vertical-rl/image-inside-nested-blocks-with-border-expected.txt

ALERT: 210 300
FAIL: 1 cases failed
